Exception handling -delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -3,12 +3,14 @@ package com.educandoweb.course.services;
 import java.util.List;
 import java.util.Optional;
 
-import org.apache.catalina.startup.ClassLoaderFactory.Repository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DataBaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -40,9 +42,19 @@ public class UserService {
 	
 	public void delete(Long id) {
 		
-		userRepository.deleteById(id);
+		if(!userRepository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
 		
- }
+		try {
+			userRepository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
+	}
+		
+		
+ 
 	
 	public User update(Long id,User obj) {
 		User entity = userRepository.getReferenceById(id);// getReferenceById() Proxy (casca vazia com ID)
